Struct wcxhead::PackDefaultParamStruct

#[repr(C)]
pub struct PackDefaultParamStruct {
    pub size: c_int,
    pub PluginInterfaceVersionLow: DWORD,
    pub PluginInterfaceVersionHi: DWORD,
    pub DefaultIniName: [c_char; 260],
}

PackDefaultParamStruct is passed to PackSetDefaultParams to inform the plugin about the current plugin interface version and ini file location.

Declaration:

typedef struct {

int size;
    DWORD PluginInterfaceVersionLow;
    DWORD PluginInterfaceVersionHi;
    char DefaultIniName[MAX_PATH];
} PackDefaultParamStruct;

Description of struct members:

size 	The size of the structure, in bytes. Later revisions of the plugin interface may add more structure members, and will adjust this size field accordingly.

PluginInterfaceVersionLow 	Low value of plugin interface version. This is the value after the comma, multiplied by 100! Example. For plugin interface version 2.1, the low DWORD is 10 and the high DWORD is 2..

PluginInterfaceVersionHi 	High value of plugin interface version.

DefaultIniName 	Suggested location+name of the ini file where the plugin could store its data. This is a fully qualified path+file name, and will be in the same directory as the wincmd.ini. It's recommended to store the plugin data in this file or at least in this directory, because the plugin directory or the Windows directory may not be writable!.

Fields
size: c_intPluginInterfaceVersionLow: DWORDPluginInterfaceVersionHi: DWORDDefaultIniName: [c_char; 260]